requirements.txt
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
et_xmlfile==2.0.0
filelock==3.15.4
Flask==3.0.3
Flask-Excel==0.0.7
Flask-Session==0.8.0
gTTS==2.5.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
lml==0.1.0
MarkupSafe==2.1.5
msgspec==0.18.6
mysql-connector-python==9.1.0
openpyxl==3.1.5
platformdirs==4.2.2
playsound==1.2.2
PyAudio==0.2.14
pyexcel==0.7.1
pyexcel-io==0.6.7
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
requests==2.32.3
SpeechRecognition==3.10.4
texttable==1.7.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.4
